+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix `Network::delete` endpoint
 - Fix `Network::disconnect` endpoint. It now uses correct `ContainerDisconnectionOpts`.
 - Add `ContainerCreateOptsBuilder::security_opts` function to specify container security options [#12](https://github.com/vv9k/docker-api-rs/pull/12)
+- Add `NetworkCreateOptsBuilder::ipam` function to specify ip address management config of a network
 
 # 0.7.0
 - Make `PidsStats`, `Descriptor` and `DistributionInspectInfo` struct fields publicly accessible.